new docker image because of problem with Selenium python version changed to 3 10 dut to those problems bot runs easy. All's good.
