+++ Makefile
@@ -49,7 +49,7 @@ help:
 
 
 
-# Создает шаблонный файл .pot из app.py (просканирует например app.py и найдёт места для переводов)
+# Создает шаблонный файл .pot из main.py (просканирует например main.py и найдёт места для переводов)
 create_for_app_translate:
 	xgettext -d messages -o src/middleware/locales/messages.pot src/app/handlers.py --from-code UTF-8
