+++ pyproject.toml
@@ -8,7 +8,7 @@ readme = "README.md"
 
 
 [tool.poetry.dependencies]
-python = "^3.12"
+python = "^3.10"
 sqlalchemy = "^2.0.38"
 greenlet = "^3.1.1"
 asyncpg = "^0.30.0"
@@ -21,6 +21,8 @@ aioschedule = "^0.5.2"
 schedule = "^1.2.2"
 fastapi = "^0.115.9"
 uvicorn = "^0.34.0"
+chromedriver-binary = "^135.0.7039.0.0"
+pyyaml = "^6.0.2"
 
 
 [tool.poetry.group.dev.dependencies]
