--- src/bot_instance.py
@@ -1,16 +1,0 @@
-from aiogram import Bot, Dispatcher
-from aiogram.client.default import DefaultBotProperties
-from aiogram.enums import ParseMode
-from src.settings import settings
-
-TOKEN = settings.TOKEN
-
-
-dp = Dispatcher()
-bot = Bot(token=TOKEN, default=DefaultBotProperties(parse_mode=ParseMode.HTML))
-
-all = ['bot']
-# dp.include_routers(
-#     start_router,
-#     echo_router,
-# )
