--- src/src/Exam/time_constants.py
@@ -1,4 +1,0 @@
-constant_text_ten_mins = "Внимание! Прошло 10 минут. Осталось 5 минут"
-constant_text_five_mins = "Внимание! Прошло 5 минут. Осталось 10 минут"
-constant_text_thirteen_mins = "Внимание! Прошло 13 минут. Осталось 2 минуты"
-constant_text_time_has_passed = "Внимание! Время закончилось."
