--- src/src/selenium_parcer/selenium_instance.py
@@ -1,11 +1,0 @@
-from selenium import webdriver
-from selenium.webdriver.chrome.service import Service
-from webdriver_manager.chrome import ChromeDriverManager
-
-
-options = webdriver.ChromeOptions()
-options.add_argument("--disable-blink-features=AutomationControlled")
-options.add_argument("--window-size=1920,1080")
-
-driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()), options=options)
-
