--- src/src/selenium_parcer/utils.py
@@ -1,1 +1,0 @@
-#TODO switch here yml file parcer for links
